HU003 Terminado

+++ src/main/java/ApiRest/TurnosRotativos/repository/EmpleadoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmpleadoRepository extends JpaRepository<Empleado, Long> {
+public interface EmpleadoRepository extends JpaRepository<Empleado, Integer> {
 
     boolean existsByNroDocumento(int nroDocumento);
     boolean existsByEmail(String email);
